--- a/ultralytics/data/utils.py
+++ b/ultralytics/data/utils.py
@@ -49,10 +49,16 @@ def img2label_paths(img_paths):
 
 def get_hash(paths):
     """Returns a single hash value of a list of paths (files or dirs)."""
-    size = sum(os.path.getsize(p) for p in paths if os.path.exists(p))  # sizes
-    h = hashlib.sha256(str(size).encode())  # hash sizes
-    h.update("".join(paths).encode())  # hash paths
-    return h.hexdigest()  # return hash
+    hasher = hashlib.sha256()
+    for path in paths:
+        try:
+            stat_info = os.stat(path)
+            # file size and modification time
+            file_info = f"{stat_info.st_size}|{stat_info.st_mtime}"
+            hasher.update(file_info.encode())
+        except (OSError, IOError):
+            continue
+    return hasher.hexdigest()
 
 
 def exif_size(img: Image.Image):