--- a/ultralytics/data/utils.py
+++ b/ultralytics/data/utils.py
@@ -50,14 +50,16 @@ def img2label_paths(img_paths):
 def get_hash(paths):
     """Returns a single hash value of a list of paths (files or dirs)."""
     hasher = hashlib.sha256()
+    stat_sum = 0
     for path in paths:
         try:
             stat_info = os.stat(path)
             # file size and modification time
-            file_info = f"{stat_info.st_size}|{stat_info.st_mtime}"
-            hasher.update(file_info.encode())
+            stat_sum = stat_info.st_size + int(stat_info.st_mtime * 1000)
         except OSError:
             continue
+    hasher.update(''.join(paths).encode())  # hash paths
+    hasher.update(str(stat_sum))  # hash file size and modification time
     return hasher.hexdigest()