--- a/ultralytics/data/utils.py
+++ b/ultralytics/data/utils.py
@@ -52,9 +52,9 @@ def get_hash(paths):
     stat_sum = 0
     for path in paths:
         stat_info = os.stat(path)
-        stat_sum += stat_info.st_size + int(stat_info.st_mtime * 1000)  # file size and modification time
-    h = hashlib.sha256("".join(paths).encode())  # hash paths
-    h.update(str(stat_sum).encode())  # hash file size and modification time
+        stat_sum += stat_info.st_size + int(stat_info.st_mtime * 1000)
+    h = hashlib.sha256(str(stat_sum).encode())  # hash file size and modification time
+    h.update("".join(paths).encode())   # paths
     return h.hexdigest()