Claude suggested `os.stat` and `try..except` for faster hashing.

**main**
```
%time get_hash(Path("/datasets/coco/images/train2017/").glob('*'))
CPU times: user 1.31 s, sys: 2.44 s, total: 3.75 s
Wall time: 3.75 s
```
**PR**
```
In [38]: %time get_hash(Path("/datasets/coco/images/train2017/").glob('*'))
CPU times: user 1.05 s, sys: 1.28 s, total: 2.33 s
Wall time: 2.33 s
```


<!--
Thank you 🙏 for your contribution to [Ultralytics](https://www.ultralytics.com/) 🚀! Your effort in enhancing our repositories is greatly appreciated. To streamline the process and assist us in integrating your Pull Request (PR) effectively, please follow these steps:

1. Check for Existing Contributions: Before submitting, kindly explore existing PRs to ensure your contribution is unique and complementary.
2. Link Related Issues: If your PR addresses an open issue, please link it in your submission. This helps us better understand the context and impact of your contribution.
3. Elaborate Your Changes: Clearly articulate the purpose of your PR. Whether it's a bug fix or a new feature, a detailed description aids in a smoother integration process.
4. Ultralytics Contributor License Agreement (CLA): To uphold the quality and integrity of our project, we require all contributors to sign the CLA. Please confirm your agreement by commenting below:

    I have read the CLA Document and I sign the CLA

For more detailed guidance and best practices on contributing, refer to our ✅ [Contributing Guide](https://docs.ultralytics.com/help/contributing/). Your adherence to these guidelines ensures a faster and more effective review process.
-->


## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary  
Improved the way file hashes are generated for better accuracy and reliability. 🔒✨

### 📊 Key Changes  
- Updated the `get_hash` function to use both file size and modification time for hashing, instead of just file size and path.
- Improved error handling to skip files that can't be accessed, making the process more robust.

### 🎯 Purpose & Impact  
- Ensures more accurate detection of file changes, reducing the risk of missing updates.
- Makes dataset and file management more reliable, especially when files are updated or modified.
- Enhances stability by gracefully handling inaccessible files, preventing crashes during hashing.