@UltralyticsAssistant 2025-04-16
👋 Hello @Y-T-G, thank you for submitting an `ultralytics/ultralytics` 🚀 PR! This is an automated response to help streamline the review process—an Ultralytics engineer will assist you soon.

Please review the following checklist to ensure your contribution can be integrated smoothly:

- ✅ **Define a Purpose**: Clearly explain the purpose of your fix or feature in your PR description, and link to any [relevant issues](https://github.com/ultralytics/ultralytics/issues). Ensure your commit messages are clear, concise, and follow the project's conventions.
- ✅ **Synchronize with Source**: Make sure your PR is up-to-date with the `ultralytics/ultralytics` `main` branch. If it's behind, update it via the 'Update branch' button or by running `git pull` and `git merge main` locally.
- ✅ **Ensure CI Checks Pass**: Check that all Ultralytics [Continuous Integration (CI)](https://docs.ultralytics.com/help/CI/) checks are passing. If any fail, please resolve them.
- ✅ **Update Documentation**: Update the relevant [documentation](https://docs.ultralytics.com/) for any new or modified features.
- ✅ **Add Tests**: If applicable, include or update tests to cover your changes, and confirm all tests are passing.
- ✅ **Sign the CLA**: If this is your first Ultralytics PR, please sign our [Contributor License Agreement (CLA)](https://docs.ultralytics.com/help/CLA/) by writing "I have read the CLA Document and I sign the CLA" in a new comment.
- ✅ **Minimize Changes**: Limit your changes to the **minimum** required for your bug fix or feature addition. _"It is not daily increase but daily decrease, hack away the unessential. The closer to the source, the less wastage there is."_ — Bruce Lee

For further details, please refer to our [Contributing Guide](https://docs.ultralytics.com/help/contributing/). If you have any questions, feel free to leave a comment. Thank you for helping improve Ultralytics! 🌟🦾