@UltralyticsAssistant 2025-04-16
🎉 Fantastic work @Y-T-G and @glenn-jocher—your improvements to file hashing in our data utilities have just been merged! Your attention to robust error handling truly embodies Henry Ford’s words: “Quality means doing it right when no one is looking.” This update not only makes our data processing more reliable but also enhances the experience for everyone working with Ultralytics HUB and beyond. Thank you for your dedication and thoughtful contributions—together, we’re setting new standards for excellence! 🚀