@Y-T-G 2025-04-16
@glenn-jocher Not sure. The try except is just a more efficient alternative of `os.path.exists()` that was already present. Don't know if checking is required at all.